Add more control to manual mode
Manual mode forces tester to run each test individually by
clicking buttons in the browser. Also, when a test calls iso.report()
in manual mode, rather than returning to the testing home screen, the
test result is shown at the bottom of the reporting page along with a
button that will return the tester to the home screen. This way, a
tester can inspect console-messages/DOM/whatever for errors before
finishing the test.